towards a repl
we want more than one command, so let's move things into a loop. You'll need to use a keyboard signal to break out of this.

rust flushes stdout when it hits a newline, so println and print have a slightly different behavior and we need to force a fush to create a prod.