add basic pipes support
this is a really clunky way of implementing pipes. We introduce an outer loop looking for pipes, connect together stdin and out where necessary, and run our old inner loop on each sewn together segment.

This isn't super elegant but it does work for the basic use of pipes. I'm not sure I'd trust this with running Unicorn, though.